functions/python/requirements.txt
firebase-functions
firebase-admin
python-docx
google-cloud-storage
google-cloud-firestore